fix: update forklift pallet for v2 config changes
