+++ forklift/requirements.txt
@@ -1,2 +1,1 @@
-pygsheets==1.1.4
-oauth2client==4.1.3
+pygsheets==2.0.6
