+++ forklift/settings/dev.py
@@ -2,11 +2,8 @@
 import os
 
 # URLS
-queryLayersUrl = 'https://docs.google.com/spreadsheets/d/1aVJ68hOyp4H6sKEEuL-xtB2qE_y6W0gDg35TgUSxtFg/edit#gid=0'
-mapServiceJson = 'http://localhost/arcgis/rest/services/DEQEnviro/MapService/MapServer?f=json'
-securedServiceJson = 'http://localhost/arcgis/rest/services/DEQEnviro/Secure/MapServer?f=json'
-webdata = r'W:\deq-enviro\_src\webdata'
 agsServer = 'localhost'
+configSpreadsheetUrl = 'https://docs.google.com/spreadsheets/d/1aVJ68hOyp4H6sKEEuL-xtB2qE_y6W0gDg35TgUSxtFg/edit#gid=0'
 
 # database connections
 dbConnects = os.path.join(os.path.dirname(os.path.abspath(__file__)), r'..\databases')
@@ -20,9 +17,6 @@ fgd = 'deqquerylayers.gdb'
 sendEmails = False
 
 reportEmail = 'stdavis@utah.gov'
-
-mapData1 = r'C:\MapData'
-mapData2 = r'C:\MapData'
 
 #: these layers cause issues for forklift so they are skipped and manually updated within the pallet
 PROBLEM_LAYERS = ['DEQMAP_EIChemical']
