+++ forklift/settings/prod.py
@@ -4,7 +4,7 @@ from .stage import *
 sgid = {'ENVIRONMENT': os.path.join(dbConnects, 'SGID as ENVIRONMENT on prod.sde'),
         'WATER': os.path.join(dbConnects, 'SGID as WATER on prod.sde'),
         'ENERGY': os.path.join(dbConnects, 'SGID as ENERGY on prod.sde')}
-queryLayersUrl = 'https://docs.google.com/spreadsheets/d/1-HJViZduddtnxSq35yzhR6YU0U6Pwo0OaimXPHT4vPw/edit#gid=0'
+configSpreadsheetUrl = 'https://docs.google.com/spreadsheets/d/1-HJViZduddtnxSq35yzhR6YU0U6Pwo0OaimXPHT4vPw/edit#gid=0'
 utahddw_file_name = 'utahddw.sde'
 
 sendEmails = True
